+++ index.html
@@ -23,7 +23,7 @@
     -->
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
 
-  <title>Classic Cuts | Home</title>
+  <title>Home | Classic Cuts</title>
 </head>
 
 <body>
