+++ src/components/checkout/ShippingDetailsForm.jsx
@@ -1,0 +1,152 @@
+import React, { useState, useContext, useEffect } from "react";
+import TextField from "@mui/material/TextField";
+import Button from "@mui/material/Button";
+import { FormControl, Grid } from "@mui/material";
+import { CheckoutStepContext } from "../../context/CheckoutStep/Step";
+
+export function ShippingDetailsForm({ onNext }) {
+  const newCheckoutStepContext = useContext(CheckoutStepContext);
+
+  const [formData, setFormData] = useState({
+    firstName: "",
+    lastName: "",
+    addressLine1: "",
+    addressLine2: "",
+    city: "",
+    state: "",
+    country: "",
+    zipCode: "",
+  });
+
+  const handleChange = (e) => {
+    const { name, value } = e.target;
+    setFormData({
+      ...formData,
+      [name]: value,
+    });
+  };
+
+  const handleNext = (e) => {
+    e.preventDefault();
+    newCheckoutStepContext.setStepValue(newCheckoutStepContext.stepValue + 1);
+    onNext(formData);
+  };
+
+  return (
+    <>
+      <label className="block font-medium ml-0 m-3">Shipping details</label>
+      <form onSubmit={handleNext} className="form-no-outline m-3">
+        <Grid container spacing={3}>
+          <Grid item xs={12} sm={6}>
+            <FormControl fullWidth variant="outlined">
+              <TextField
+                fullWidth
+                label="First Name"
+                name="firstName"
+                value={formData.firstName}
+                onChange={handleChange}
+                required
+              />
+            </FormControl>
+          </Grid>
+          <Grid item xs={12} sm={6}>
+            <FormControl fullWidth variant="outlined">
+              <TextField
+                fullWidth
+                label="Last Name"
+                name="lastName"
+                value={formData.lastName}
+                onChange={handleChange}
+                required
+              />
+            </FormControl>
+          </Grid>
+          <Grid item xs={12} sm={6}>
+            <FormControl fullWidth variant="outlined">
+              <TextField
+                fullWidth
+                label="Address Line 1"
+                name="addressLine1"
+                value={formData.addressLine1}
+                onChange={handleChange}
+                required
+              />
+            </FormControl>
+          </Grid>
+          <Grid item xs={12} sm={6}>
+            <FormControl fullWidth variant="outlined">
+              <TextField
+                fullWidth
+                label="Address Line 2"
+                name="addressLine2"
+                value={formData.addressLine2}
+                onChange={handleChange}
+              />
+            </FormControl>
+          </Grid>
+          <Grid item xs={12} sm={6}>
+            <FormControl fullWidth variant="outlined">
+              <TextField
+                fullWidth
+                label="City"
+                name="city"
+                value={formData.city}
+                onChange={handleChange}
+                required
+              />
+            </FormControl>
+          </Grid>
+          <Grid item xs={12} sm={6}>
+            <FormControl fullWidth variant="outlined">
+              <TextField
+                fullWidth
+                label="State"
+                name="state"
+                value={formData.state}
+                onChange={handleChange}
+                required
+              />
+            </FormControl>
+          </Grid>
+          <Grid item xs={12} sm={6}>
+            <FormControl fullWidth variant="outlined">
+              <TextField
+                fullWidth
+                label="Country"
+                name="country"
+                value={formData.country}
+                onChange={handleChange}
+                required
+              />
+            </FormControl>
+          </Grid>
+
+          <Grid item xs={12} sm={6}>
+            <FormControl fullWidth variant="outlined">
+              <TextField
+                fullWidth
+                label="ZIP Code"
+                name="zipCode"
+                value={formData.zipCode}
+                onChange={handleChange}
+                required
+              />
+            </FormControl>
+          </Grid>
+
+          <Grid item xs={12}>
+            <Button
+              className="bg-[#9155FD] w-full"
+              type="submit"
+              variant="contained"
+              size="large"
+              sx={{ padding: ".8rem 0", bgcolor: "#9155FD" }}
+            >
+              Next
+            </Button>
+          </Grid>
+        </Grid>
+      </form>
+    </>
+  );
+}
